Moved comment layout to folder
